app/Helpers/python/script/requirements.txt
numpy==1.21.5
opencv-python==4.5.4.60
Pillow==8.4.0
protobuf==3.19.3
PyMySQL==1.0.2
